Notify user with “balloon” message when 5% of time remaining before break is reached.
I have found this feature very useful in other “break” software as it
allows the user to prepare themselves for a upcoming break. As the event
is no longer a surprise they are more likely to take the break instead
of just ignoring the software.

I have build this in as a none configurable feature but it could be
added behind a on/off toggle config if you think some users would never
want this(?).