Shown remaining break seconds in "cup"
It can sometimes be hard to read the countdown timer whilst in a break
due to the colour clash.
I have moves the seconds countdown display to the middle of the white
cup. This is more easier to see. They less eye strain the better :-)
I have also removed the text “seconds left” as I thought this was no
longer necessary.