Add user prompt to always run on startup.
On first run of this application a prompt will now be displayed asking
the user if they always want this application to run on start up.
It is very easy to forget to run apps like these and if you have
forgotten then before you know if it's too late and you have a headache
already!
Therefore I think its justified to ask to the user to auto run an app of
this kind.